+++ CHANGELOG.md
@@ -21,6 +21,8 @@ Main (unreleased)
   - `otelcol.connector.spanlogs` - creates logs from spans. It is the flow mode equivalent
   to static mode's `automatic_logging` processor. (@ptodev)
 
+  - `discovery.serverset` discovers Serversets stored in Zookeeper. (@thampiotr)
+
   - `discovery.triton` discovers scrape targets from Triton Container Monitor. (@erikbaranowski)
 
 - Flow: allow the HTTP server to be configured with TLS in the config file